simplified and standardized code & arrows
